chore: rename dynamo_snapshot.py to snapshot.py
